% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemdarima.R
\name{eemdarima}
\alias{eemdarima}
\title{Function to predict with EEMD-ARIMA model}
\usage{
eemdarima(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EEMD-ARIMA model
}
\description{
Function to predict with EEMD-ARIMA model
}
\examples{
# eemdarima(data = nottem, n.ahead = 6)
}

