% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemdpsf.R
\name{eemdpsf}
\alias{eemdpsf}
\title{Function to predict with EEMD-PSF model}
\usage{
eemdpsf(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EEMD-PSF model
}
\description{
Function to predict with EEMD-PSF model
}
\examples{
# eemdpsf(data = nottem, n.ahead = 6)
}
