% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemdpsfarima.R
\name{eemdpsfarima}
\alias{eemdpsfarima}
\title{Function to predict with EEMD-PSF,ARIMA model}
\usage{
eemdpsfarima(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EEMD-PSF,ARIMA model
}
\description{
Function to predict with EEMD-PSF,ARIMA model
}
\examples{
# eemdpsfarima(data = nottem, n.ahead = 6)
}

