% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdarima.R
\name{emdarima}
\alias{emdarima}
\title{Function to predict with EMD-ARIMA model}
\usage{
emdarima(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EMD-ARIMA model
}
\description{
Function to predict with EMD-ARIMA model
}
\examples{
# emdarima(data = nottem, n.ahead = 6)
}

