% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdpsf.R
\name{emdpsf}
\alias{emdpsf}
\title{Function to predict with EMD-PSF model}
\usage{
emdpsf(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EMD-PSF model
}
\description{
Function to predict with EMD-PSF model
}
\examples{
# emdpsf(data = nottem, n.ahead = 6)
}

