% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdpsfarima.R
\name{emdpsfarima}
\alias{emdpsfarima}
\title{Function to predict with EMD-PSF,ARIMA model}
\usage{
emdpsfarima(data, n.ahead)
}
\arguments{
\item{data}{as input time series data}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
predicted values with EMD-PSF,ARIMA model
}
\description{
Function to predict with EMD-PSF,ARIMA model
}
\examples{
# emdpsfarima(data = nottem, n.ahead = 6)
}

