% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsf.R
\name{lpsf}
\alias{lpsf}
\title{Function to restrict the legth of dataset in multiples of 24}
\usage{
lpsf(data, n.ahead)
}
\arguments{
\item{data}{as inpute time series}

\item{n.ahead}{as horizon of values to be predicted}
}
\value{
returns the predictied results
}
\description{
Function to restrict the legth of dataset in multiples of 24
}
